+++ Mods ChoGGi/Game Rules Enable Achievements/metadata.lua
@@ -1,0 +1,20 @@
+return PlaceObj("ModDef", {
+	"title", "Game Rules Enable Achievements",
+	"id", "ChoGGi_GameRulesEnableAchievements",
+	"steam_id", "2707063636",
+	"pops_any_uuid", "a5053041-f913-4dcf-a2fb-309522cb6024",
+	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.png",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagGameplay", true,
+	"description", [[
+Certain game rules will no longer disable achievements.
+]],
+})